Move emitter helper dialog related events to main
